test: 여행 계획 삭제 API docs 테스트 코드 구현

+++ src/main/java/com/travelsketch/travel/api/controller/trip/request/ModifyTripRequest.java
@@ -11,10 +11,10 @@ import java.util.List;
 public class ModifyTripRequest {
 
     private String title;
-    private List<Long> attractions;
+    private List<Integer> attractions;
 
     @Builder
-    private ModifyTripRequest(String title, List<Long> attractions) {
+    private ModifyTripRequest(String title, List<Integer> attractions) {
         this.title = title;
         this.attractions = attractions;
     }
